requirements.txt
roboflow 
ultralytics
fastapi[all]
torch
torchvision
opencv-python
pillow
celery[redis]
jinja2
python-multipart